realtek: provide phy local SoC info structure
Until now the Realtek targets have a global SoC info structure that is filled
during kernel bringup. There has always been the discussion to remove it but
it was never implemented. For the PHY driver this structure is mostly inadequate
and only used in a few places.

Create a new local SoC info structure that serves the drivers needs and fill
it during module initialization. To modernize the SoC register access make use
of regmap for the new functions. For that lookup the memory addresses from
the only DTS node that has compatible "syscon" set. This is currently always
switchcore@1b000000 with our desired memory range reg = <0x1b000000 0x10000>.

Signed-off-by: Markus Stockhausen <markus.stockhausen@gmx.de>